chore(deps): bump all dependencies to latest via cargo update
Bumped workspace constraints and waf-cli deps, ran  to resolve.
Cargo update resolved everything — no further bumps possible.

Version bumps (constraint → locked version):
- axum 0.7 → 0.8 (locked 0.8.9)
- tower 0.4 → 0.5 (locked 0.5.3)
- tower-http 0.5 → 0.6 (locked 0.6.10)
- toml 0.8 → 0.9 (locked 0.9.12)
- rusqlite 0.37 → 0.39 (locked 0.39.0)
- schemars 0.8 → 0.9 (locked 0.9.0)
- reqwest 0.12 (waf-cli, locked 0.12.28)
- rcgen 0.12 → 0.14 (waf-cli, locked 0.14.7)
- tokio-rustls 0.25 → 0.26 (waf-cli, locked 0.26.4)
- rustls-pemfile 2.0 → 2 (locked 2.2.0)
- prometheus-client 0.22 → 0 (locked 0.24.1)
- maxminddb 0.24 → 0.28 (locked 0.28.1)
- clap/clap_complete 4.5 → 4 (locked 4.6.x)

API fixes for dependency upgrades:
- schemars 0.9: schema_for!() returns Schema directly, no .schema field
- rusqlite 0.39: COUNT(*) returns i64 (not u64), params must be i64 (not usize)

+++ waf-core/src/database.rs
@@ -136,7 +136,7 @@ impl WafDatabase {
     }
 
     /// Get recent threats (last N entries).
-    pub fn get_recent_threats(&self, limit: usize) -> Result<Vec<ThreatEntry>> {
+    pub fn get_recent_threats(&self, limit: i64) -> Result<Vec<ThreatEntry>> {
         let mut stmt = self.conn.prepare(
             "SELECT ip, path, threat_score, action, rule_ids, timestamp FROM audit_log ORDER BY timestamp DESC LIMIT ?1"
         )?;
@@ -165,7 +165,7 @@ impl WafDatabase {
         self.conn.query_row(
             "SELECT COUNT(*) FROM user_whitelist WHERE ip = ?1",
             [ip],
-            |row| row.get::<_, u64>(0),
+            |row| row.get::<_, i64>(0),
         )
         .map(|count| count > 0)
         .unwrap_or(false)
@@ -186,11 +186,11 @@ impl WafDatabase {
 
     // ── Stats ───────────────────────────────────────────
 
-    pub fn get_blocklist_count(&self) -> Result<u64> {
+    pub fn get_blocklist_count(&self) -> Result<i64> {
         self.conn.query_row("SELECT COUNT(*) FROM blocklist_ips", [], |row| row.get(0))
     }
 
-    pub fn get_audit_count(&self) -> Result<u64> {
+    pub fn get_audit_count(&self) -> Result<i64> {
         self.conn.query_row("SELECT COUNT(*) FROM audit_log", [], |row| row.get(0))
     }
 }
